+++ event-sourcing/event-repository/event-repository-jdbc/src/main/java/uk/gov/justice/services/eventsourcing/repository/jdbc/eventlog/EventLogConverter.java
@@ -1,7 +1,7 @@
 package uk.gov.justice.services.eventsourcing.repository.jdbc.eventlog;
 
 import uk.gov.justice.services.common.converter.StringToJsonObjectConverter;
-import uk.gov.justice.services.common.util.DateTimeProvider;
+import uk.gov.justice.services.common.util.Clock;
 import uk.gov.justice.services.eventsourcing.common.exception.InvalidStreamIdException;
 import uk.gov.justice.services.messaging.DefaultJsonEnvelope;
 import uk.gov.justice.services.messaging.JsonEnvelope;
@@ -28,7 +28,7 @@ public class EventLogConverter {
     StringToJsonObjectConverter stringToJsonObjectConverter;
 
     @Inject
-    DateTimeProvider dateTimeProvider;
+    Clock clock;
 
     /**
      * Creates an {@link EventLog} object from the <code>eventEnvelope</code>.
@@ -52,7 +52,7 @@ public class EventLogConverter {
                 eventMetadata.name(),
                 envelope.metadata().asJsonObject().toString(),
                 extractPayloadAsString(envelope),
-                dateTimeProvider.now());
+                clock.now());
     }
 
     /**